# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Plant Control Monitoring API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/term/{id_pemilik_lahan}": {
            "get": {
                "tags": [
                    "Term"
                ],
                "summary": "get list term",
                "operationId": "getTerms",
                "parameters": [
                    {
                        "name": "id_pemilik_lahan",
                        "in": "path",
                        "description": "you need login first before using this api",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/laporan": {
            "post": {
                "tags": [
                    "Laporan"
                ],
                "summary": "post laporan",
                "operationId": "laporanPost",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name Laporan",
                        "type": "string"
                    },
                    {
                        "name": "note",
                        "in": "formData",
                        "description": "Note from farmmanager",
                        "type": "string"
                    },
                    {
                        "name": "varietas",
                        "in": "formData",
                        "description": "varietas tanaman",
                        "type": "string"
                    },
                    {
                        "name": "is_overdue",
                        "in": "formData",
                        "description": "is overdue",
                        "type": "integer"
                    },
                    {
                        "name": "foto1",
                        "in": "formData",
                        "description": "optional foto1",
                        "type": "file"
                    },
                    {
                        "name": "foto2",
                        "in": "formData",
                        "description": "optional foto2",
                        "type": "file"
                    },
                    {
                        "name": "foto3",
                        "in": "formData",
                        "description": "optional foto3",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/laporans": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "get list laporan by pemilik lahan",
                "operationId": "getlistlaporan",
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/laporan/{id_laporan}": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "get detaillaporan by id laporan",
                "operationId": "getdetaillaporanbyid",
                "parameters": [
                    {
                        "name": "id_laporan",
                        "in": "path",
                        "description": "id laporan",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/task": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "post task",
                "operationId": "posttask",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "post task",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "id_pemilik_lahan": {
                                    "type": "integer"
                                },
                                "id_term": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}